pyproject.toml
[tool.poetry]
name = "revobank-api"
version = "0.1.0"
description = "RESTful API for RevoBank application"
authors = ["Rifqi <rifqiramadani@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.11.1"
flask = "^2.3.0"
pyjwt = "^2.6.0"

[tool.poetry.dev-dependencies]
pytest = "^7.0.0"
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[project]
name = "revobank-api"
version = "0.1.0"
description = "RESTful API for RevoBank application"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "--cov=."